Relocate types.py and dev.py from util to _core to prevent MPI init on loading simsopt
