+++ src/simsopt/_core/__init__.py
@@ -1,8 +1,6 @@
 from .derivative import *
 from .optimizable import *
-from .finite_difference import *
 from .util import *
 
-__all__ = (derivative.__all__ + optimizable.__all__ +
-           finite_difference.__all__ + util.__all__)
+__all__ = (derivative.__all__ + optimizable.__all__ + util.__all__)
