+++ src/simsopt/objectives/least_squares.py
@@ -19,14 +19,13 @@ import numpy as np
 
 from .._core.optimizable import Optimizable
 from .._core.util import ObjectiveFailure
-from ..util.types import RealArray, IntArray, BoolArray
+from .._core.types import RealArray, IntArray, BoolArray
 
+__all__ = ['LeastSquaresProblem']
 
 logger = logging.getLogger(__name__)
 
 StrSeq = Union[Sequence, Sequence[Sequence[str]]]
-
-__all__ = ['LeastSquaresProblem']
 
 
 class LeastSquaresProblem(Optimizable):
