+++ src/simsopt/solve/mpi.py
@@ -23,8 +23,6 @@ except ImportError as err:
 
 from .._core.optimizable import Optimizable
 from ..util.mpi import MpiPartition
-from ..util.types import RealArray
-from .._core.util import finite_difference_steps
 from .._core.finite_difference import MPIFiniteDifference
 from ..objectives.least_squares import LeastSquaresProblem
